+++ BaloonsPop/MainMenu.cs
@@ -16,9 +16,9 @@
             logger.Log("|____||_| |_||____||____||_____||_____||_|   |_|   |_|    |_____||_|    |_____|  ");
             logger.Log("");
             logger.Log("                    Welcome to “Balloons Pops” game :)");
-            logger.Log("");  
+            logger.Log("");
             logger.Log("                     Please try to pop the balloons!");
-            logger.Log("");   
+            logger.Log("");
             logger.Log("                    Press \"Enter\" to start the game!");
             Console.ReadLine();
             Console.Clear();
@@ -27,7 +27,7 @@
             logger.Log("* Use 'restart' to start a new game!");
             logger.Log("* Use 'exit' to quit the game!");
             logger.Log("");
-            
+
             Game game = new Game(logger);
 
             while (true)